changed folder structure for clarity
