+++ domain/src/AggregateRootId.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Domain\Aggregates;
+namespace Domain;
 
 use EventSauce\EventSourcing\AggregateRootId as AggregateRootIdInterface;
 use Illuminate\Support\Str;
