+++ domain/src/Nft/Actions/AcquireNft.php
@@ -1,15 +1,15 @@
 <?php
 
-namespace Domain\Actions;
+namespace Domain\Nft\Actions;
 
-use Domain\Aggregates\NftId;
+use Domain\Nft\NftId;
 use Domain\ValueObjects\FiatAmount;
 
-final class DisposeOfNft
+final class AcquireNft
 {
     public function __construct(
         public readonly NftId $nftId,
-        public readonly FiatAmount $disposalProceeds,
+        public readonly FiatAmount $costBasis,
     ) {
     }
 }
