+++ domain/src/Nft/NftId.php
@@ -1,6 +1,8 @@
 <?php
 
-namespace Domain\Aggregates;
+namespace Domain\Nft;
+
+use Domain\AggregateRootId;
 
 final class NftId extends AggregateRootId
 {
